Core/change db lib

+++ core/core-cart/src/commonMain/kotlin/com/foodsaver/app/domain/model/CartItemModel.kt
@@ -6,5 +6,5 @@ data class CartItemModel(
     val localId: Long,
     val globalId: String?,
     val product: ProductModel,
-    val quantity: Int
+    val quantity: Long
 )
